#23 - responsive search area with adding and deleting
